downdetector_app/requirements.txt
aws-cdk-lib==2.206.0
constructs>=10.0.0,<11.0.0